+++ ngx-numeric-range-form-field/projects/ngx-numeric-range-form-field/src/public-api.ts
@@ -4,3 +4,4 @@
 
 export * from './lib/ngx-numeric-range-form-field.module';
 export { INumericRange } from './lib/form/model/numeric-range-field.model';
+export { NumericRangeFormFieldContainerComponent } from './lib/container/numeric-range-form-field-container.component';